Refactor: create shared view model for watchlist and favorite
